requirements.txt
Django>=4.2
django-allauth
django-ckeditor
Pillow
psycopg2-binary